Created the function used to sign the SQRL url + added some comments for clarity
The function uses the derived per site private key to create a signature that will go along with the http request to the sqrl server. the server should be able to verify the signature using the paired public key that also gets sent in the request. Added some comments for my own sanity and to help understand the code